created routes but want to think more about how to use cookies in the requests
