+++ controllers/eventController.js
@@ -4,9 +4,22 @@ module.exports = {
     create: (req, res) => {
         db.Event
             .create(req.body, {})
-            .then(event => 
-                res.send(event.id))
+            .then(event => {
+                res.json(event)
+            })
             .catch(err => res.json(err))
+    },
+
+    delete: (req,res) => {
+        db.Event
+            .destroy({
+                where: {id: req.body.eventId}
+            })
+            .then(event => {
+                res.json(event)
+            })
+            .catch( err => res.json(err))
+
     }
 
 }
