+++ routes/api/event.js
@@ -6,5 +6,12 @@ router
     .route("/")
     .post(eventController.create)
 
+//Matches with "/api/event/eventId"
+router
+    .route("/eventId")
+    .delete(eventController.delete)
+
+
+
 
 module.exports = router;
